+++ src/main/java/ru/yandex/practicum/filmorate/dao/Storage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.dao;
 
 import ru.yandex.practicum.filmorate.exception.ValidationException;
