+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/AbstractStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.dao.impl;
 
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
