+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/InMemoryUserStorage.java
@@ -1,8 +1,10 @@
-package ru.yandex.practicum.filmorate.storage;
+package ru.yandex.practicum.filmorate.dao.impl;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.dao.UserStorage;
+import ru.yandex.practicum.filmorate.dao.impl.AbstractStorage;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
@@ -11,7 +13,7 @@ import java.util.List;
 @Component
 @Slf4j
 @Qualifier("InMemoryUserStorage")
-public class InMemoryUserStorage extends AbstractStorage<User> implements UserStorage{
+public class InMemoryUserStorage extends AbstractStorage<User> implements UserStorage {
     private Integer usersId = 0;
 
     @Override
