+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/ReviewDbStorage.java
@@ -14,7 +14,7 @@ import ru.yandex.practicum.filmorate.exception.IncorrectParameterException;
 import ru.yandex.practicum.filmorate.exception.ObjectNotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Review;
-import ru.yandex.practicum.filmorate.storage.ReviewStorage;
+import ru.yandex.practicum.filmorate.dao.ReviewStorage;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
